refactor(symbolic): split expression module
