+++ crates/evm/symbolic/src/runtime.rs
@@ -6,8 +6,7 @@ mod calldata;
 mod cheatcodes;
 mod control;
 mod evm;
-mod expressions;
-mod hashcons;
+mod expr;
 mod memory;
 mod precompiles;
 mod solver;
@@ -20,7 +19,7 @@ pub(crate) use calldata::*;
 pub(crate) use cheatcodes::*;
 pub(crate) use control::*;
 pub(crate) use evm::*;
-pub(crate) use expressions::*;
+pub(crate) use expr::*;
 pub(crate) use memory::*;
 pub(crate) use precompiles::*;
 pub use solver::PortfolioDiagnostics;
